requirements.txt
Django==1.6.1
South==0.8.4
django-model-utils==2.0.1
mimeparse==0.1.3
python-dateutil==2.2
pytz==2013.9
six==1.5.2
wsgiref==0.1.2
